% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subitrair.R
\name{subitrair}
\alias{subitrair}
\title{Function for subtrac}
\usage{
subitrair(a, b)
}
\arguments{
\item{a}{One number}

\item{b}{Another Number}
}
\description{
This Function works for subtract 2 numbers
}
\examples{
sub(3,4)



}
